refactor: update component imports for unified module paths and improve server initialization error handling
