+++ src/app/admin/layout.tsx
@@ -7,7 +7,7 @@ import { usePathname, useRouter } from 'next/navigation'
 import { useEffect, useState, type PropsWithChildren } from 'react'
 import { isNumber } from 'remeda'
 import { ADMIN_ROUTES } from '@/app/admin/config/routes'
-import { LoadingSpinner } from '@/components/ui/LoadingSpinner'
+import { LoadingSpinner } from '@/components/ui'
 import { api } from '@/lib/api'
 import { cn } from '@/lib/utils'
 import { hasPermission, PERMISSIONS } from '@/utils/permission-system'
